move getRepositoryUrl out of commands folder
