+++ lib/commands/publish.js
@@ -8,7 +8,7 @@ const Npm = require('../npm')
 const GitDirectory = require('../git-directory')
 const GitHub = require('../github')
 const { editMessage } = require('../editor')
-const getRepositoryUrl = require('./getRepositoryUrl')
+const getRepositoryUrl = require('../getRepositoryUrl')
 
 const debug = require('debug')
